refactor: drop support for returning errors (#21)
This PR changes the `health.CheckFunc` type and `health.Checker`
interface to disallow checks from returning errors.

To be clear, this change does not force errors to be ignored. Rather,
it forces errors to be handled at the `CheckFunc` level instead of
being handled at the `health.Checker` level.

From a health check perspective, it is not helpful having a separate
error value because an error occurred during a check should be
reflected in `health.Info` anyway. Disallowing errors to be returned
also makes it less ambiguous that checks should return a non-passing
`health.Info` when an error has occurred.